+++ jxl/src/frame/modular/buffers.rs
@@ -1,0 +1,200 @@
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
+//
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+use std::{
+    ops::DerefMut,
+    sync::atomic::{AtomicUsize, Ordering},
+};
+
+use crate::{
+    error::Result,
+    frame::modular::{ChannelInfo, IMAGE_OFFSET, IMAGE_PADDING},
+    headers::bit_depth::BitDepth,
+    image::Image,
+    util::{AtomicRefCell, AtomicRefMut},
+};
+
+use super::ModularBufferInfo;
+
+// All the information on a specific buffer needed by Modular decoding.
+#[derive(Debug)]
+pub(super) struct ModularChannel {
+    // Actual pixel buffer.
+    pub(super) data: Image<i32>,
+    // Holds additional information such as the weighted predictor's error channel's last row for
+    // the transform chunk that produced this buffer.
+    pub(super) auxiliary_data: Option<Image<i32>>,
+    // Shift of the channel (None if this is a meta-channel).
+    pub(super) shift: Option<(usize, usize)>,
+    pub(super) bit_depth: BitDepth,
+}
+
+impl ModularChannel {
+    pub fn new(size: (usize, usize), bit_depth: BitDepth) -> Result<Self> {
+        Self::new_with_shift(size, Some((0, 0)), bit_depth)
+    }
+
+    pub fn new_with_shift(
+        size: (usize, usize),
+        shift: Option<(usize, usize)>,
+        bit_depth: BitDepth,
+    ) -> Result<Self> {
+        Ok(ModularChannel {
+            data: Image::new_with_padding(size, IMAGE_OFFSET, IMAGE_PADDING)?,
+            auxiliary_data: None,
+            shift,
+            bit_depth,
+        })
+    }
+
+    fn try_clone(&self) -> Result<Self> {
+        Ok(ModularChannel {
+            data: self.data.try_clone()?,
+            auxiliary_data: self
+                .auxiliary_data
+                .as_ref()
+                .map(Image::try_clone)
+                .transpose()?,
+            shift: self.shift,
+            bit_depth: self.bit_depth,
+        })
+    }
+
+    pub fn channel_info(&self) -> ChannelInfo {
+        ChannelInfo {
+            output_channel_idx: None,
+            size: self.data.size(),
+            shift: self.shift,
+            bit_depth: self.bit_depth,
+        }
+    }
+}
+
+pub(super) const BUFFER_STATUS_NOT_RENDERED: usize = 0;
+pub(super) const BUFFER_STATUS_ZERO_FILLED: usize = 1;
+pub(super) const BUFFER_STATUS_PARTIAL_RENDER: usize = 2;
+pub(super) const BUFFER_STATUS_FINAL_RENDER: usize = 3;
+
+// Note: this type uses interior mutability to get mutable references to multiple buffers at once.
+// In principle, this is not needed, but the overhead should be minimal so using `unsafe` here is
+// probably not worth it.
+#[derive(Debug)]
+pub(super) struct ModularBuffer {
+    pub(super) data: AtomicRefCell<Option<ModularChannel>>,
+    // Number of times this buffer will be used, *including* when it is used for output.
+    pub(super) remaining_uses: AtomicUsize,
+    // Transform steps that "strongly" or "weakly" use the image data in this buffer.
+    // A "strong" usage always triggers a re-render if the image data changes.
+    // A "weak" usage only triggers a re-render if the buffer is final, or if the
+    // current re-render was not only caused by weak re-renders.
+    pub(super) used_by_transforms_strong: Vec<usize>,
+    pub(super) used_by_transforms_weak: Vec<usize>,
+    pub(super) size: (usize, usize),
+    pub(super) status: AtomicUsize,
+}
+
+impl ModularBuffer {
+    pub fn get_status(&self) -> usize {
+        self.status.load(Ordering::Relaxed)
+    }
+
+    pub fn set_status(&self, val: usize) {
+        self.status.store(val, Ordering::Relaxed);
+    }
+
+    // Iterator over (transform_id, is_strong_use)
+    pub fn users(&self, include_weak: bool) -> impl Iterator<Item = (usize, bool)> {
+        let strong = self.used_by_transforms_strong.iter().map(|x| (*x, true));
+        let weak = if include_weak {
+            &self.used_by_transforms_weak[..]
+        } else {
+            &[]
+        }
+        .iter()
+        .map(|x| (*x, false));
+        strong.chain(weak)
+    }
+
+    // Gives out a copy of the buffer + auxiliary buffer, marking the buffer as used.
+    // If this was the last usage of the buffer, does not actually copy the buffer.
+    pub fn get_buffer(&self, can_consume: bool) -> Result<ModularChannel> {
+        if !can_consume {
+            return ModularChannel::try_clone(self.data.borrow().as_ref().unwrap());
+        }
+        let mut ret = None;
+        let _ = self.remaining_uses.fetch_update(
+            Ordering::Release,
+            Ordering::Acquire,
+            |remaining_pre| {
+                let remaining = remaining_pre.checked_sub(1).unwrap();
+                if ret.is_none() {
+                    if remaining == 0 {
+                        ret = Some(Ok(self.data.borrow_mut().take().unwrap()))
+                    } else {
+                        ret = self.data.borrow().as_ref().map(ModularChannel::try_clone);
+                    }
+                } else if remaining == 0 {
+                    *self.data.borrow_mut() = None;
+                }
+                Some(remaining)
+            },
+        );
+        Ok(ret.transpose()?.unwrap())
+    }
+
+    pub fn mark_used(&self, can_consume: bool) {
+        if !can_consume {
+            return;
+        }
+        let _ = self.remaining_uses.fetch_update(
+            Ordering::Release,
+            Ordering::Acquire,
+            |remaining_pre: usize| {
+                let remaining = remaining_pre.checked_sub(1).unwrap();
+                if remaining == 0 {
+                    *self.data.borrow_mut() = None;
+                }
+                Some(remaining)
+            },
+        );
+    }
+}
+
+pub fn with_buffers<T>(
+    buffers: &[ModularBufferInfo],
+    indices: &[usize],
+    grid: usize,
+    f: impl FnOnce(Vec<&mut ModularChannel>) -> Result<T>,
+) -> Result<T> {
+    let mut bufs = vec![];
+    for i in indices {
+        // Allocate buffers if they are not present.
+        let buf = &buffers[*i];
+        let b = &buf.buffer_grid[grid];
+        let mut data = b.data.borrow_mut();
+        if data.is_none() {
+            *data = Some(ModularChannel {
+                data: Image::new_with_padding(b.size, IMAGE_OFFSET, IMAGE_PADDING)?,
+                auxiliary_data: None,
+                shift: buf.info.shift,
+                bit_depth: buf.info.bit_depth,
+            });
+        }
+
+        // Skip zero-sized *tiles*.
+        //
+        // Note that some bitstreams can contain channels with one dimension being 0 (e.g. palette
+        // meta-channel with 0 colors has size (0, 3)). Those must still participate in channel
+        // numbering (but carry no entropy-coded pixels), so we only skip when both dimensions are 0.
+        // TODO(veluca): figure out if this is the best approach or we should instead pass through
+        // empty buffers.
+        if b.size.0 == 0 && b.size.1 == 0 {
+            continue;
+        }
+
+        bufs.push(AtomicRefMut::map(data, |x| x.as_mut().unwrap()));
+    }
+    f(bufs.iter_mut().map(|x| x.deref_mut()).collect())
+}
