+++ jxl/src/frame/modular/decode/mod.rs
@@ -8,5 +8,5 @@ mod channel;
 mod common;
 mod specialized_trees;
 
-pub use bitstream::decode_modular_subbitstream;
+pub(super) use bitstream::decode_modular_subbitstream;
 pub use common::ModularStreamId;
